重构图形和ImGui渲染逻辑，添加新事件处理
删除了 GraphicsException、GraphicsWindow、RgbaFloat、SpecializationConstant 和 Viewport 类及其相关成员。
修改了 graphicsDevice.MainSwapchain.Resize 方法的参数。
删除了对 System.Numerics 等命名空间的引用，添加了对 Hexa.NET.ImGui 等命名空间的引用。
重构了 ImGuiController 类，简化了渲染和资源管理逻辑。
新增了 ImGuiRenderer 类，用于在 Vulkan 环境中渲染 ImGui 界面。
新增了多个事件处理类和方法，包括 KeyCharEventArgs、KeyEventArgs、MouseButtonEventArgs、MouseMoveEventArgs 等。
