+++ Source/Graphics/Graphics.Core/EventArgs/KeyCharEventArgs.cs
@@ -1,0 +1,6 @@
+﻿namespace Graphics.Core;
+
+public class KeyCharEventArgs(char keyChar) : EventArgs
+{
+    public char KeyChar { get; } = keyChar;
+}
