+++ Source/Graphics/Graphics.Core/EventArgs/KeyEventArgs.cs
@@ -1,0 +1,10 @@
+﻿using Silk.NET.Input;
+
+namespace Graphics.Core;
+
+public class KeyEventArgs(Key key, int scanCode) : EventArgs
+{
+    public Key Key { get; } = key;
+
+    public int ScanCode { get; } = scanCode;
+}
