+++ Source/Graphics/Graphics.Core/EventArgs/MouseButtonEventArgs.cs
@@ -1,0 +1,11 @@
+﻿using System.Numerics;
+using Silk.NET.Input;
+
+namespace Graphics.Core;
+
+public class MouseButtonEventArgs(MouseButton mouseButton, Vector2 position) : EventArgs
+{
+    public MouseButton MouseButton { get; } = mouseButton;
+
+    public Vector2 Position { get; } = position;
+}
