+++ Source/Graphics/Graphics.Core/Exceptions/GraphicsException.cs
