+++ Source/Graphics/Graphics.Core/Structs/RgbaFloat.cs
