+++ Source/Graphics/Graphics.Core/Structs/SpecializationConstant.cs
